intro
Depression and anxiety are leading causes of disability but they often remain undetected and untreated. Using smartphones and wearable devices may offer a unique source of data to detect moment by moment changes in risk factors associated with mental disorders.
AIM: The study is aimed to explore the extent to which data from smartphones could predict symptoms of depression and anxiety.